Enhance diagnostics when ptrace syscall is disabled
When ptrace(PTRACE_TRACEME, ...) is rejected by kernel with EACCES or
EPERM, it usually means that there is a kernel security policy in effect
that disallowes ptrace syscall.  Let's help users by giving them a hint
what is going on and how to deal with it.

* configure.ac: Check for libselinux.
* strace.c (show_security_policy_diagnostics): New function.
(test_ptrace_setoptions_followfork,
test_ptrace_setoptions_for_all): Use it.