Remove non-Linux code
There is a lot of non-Linux code which is known to be broken for a quite
a few years, without a single build failure fix, bug fix or even a bug
report.  We have no evidence anybody builds strace on a non-Linux
platform nowadays.  The consensus is that the non-Linux code is dead and
should be removed to make the actual code easier to understand.

References:
http://sourceforge.net/mailarchive/message.php?msg_id=28023420
http://sourceforge.net/mailarchive/message.php?msg_id=28799000

This change was made using "unifdef -B -DLINUX -Dlinux -USUNOS4 -USVR4 \
-UUNIXWARE -UFREEBSD -USUNOS4_KERNEL_ARCH_KLUDGE -UHAVE_MP_PROCFS \
-UHAVE_POLLABLE_PROCFS -UHAVE_PR_SYSCALL -UUSE_PROCFS" command.